Also update Node::deserialize with 32 bit lengths for undecrypted fields

+++ src/megaclient.cpp
@@ -18819,13 +18819,13 @@ Node *NodeManager::unserializeNode(const std::string *d, bool fromOldCache)
     if (encrypted)
     {
         // Have we encoded the node key data's length?
-        if (ptr + sizeof(uint64_t) > end)
+        if (ptr + sizeof(uint32_t) > end)
         {
             mNodes.erase(nodePosition);
             return nullptr;
         }
 
-        auto length = MemAccess::get<uint64_t>(ptr);
+        auto length = MemAccess::get<uint32_t>(ptr);
         ptr += sizeof(length);
 
         // Have we encoded the node key data?
@@ -18839,13 +18839,13 @@ Node *NodeManager::unserializeNode(const std::string *d, bool fromOldCache)
         ptr += length;
 
         // Have we encoded the length of the attribute string?
-        if (ptr + sizeof(uint64_t) > end)
+        if (ptr + sizeof(uint32_t) > end)
         {
             mNodes.erase(nodePosition);
             return nullptr;
         }
 
-        length = MemAccess::get<uint64_t>(ptr);
+        length = MemAccess::get<uint32_t>(ptr);
         ptr += sizeof(length);
 
         // Have we encoded the attribute string?
